feat(site launch): add routes + update hooks
