+++ src/services/SiteDashboardService.ts
@@ -3,7 +3,6 @@ import {
   SiteDashboardInfo,
   SiteDashboardReviewRequest,
 } from "types/siteDashboard"
-import { SiteLaunchDto } from "types/siteLaunch"
 
 import { apiService } from "./ApiService"
 
@@ -35,11 +34,4 @@ export const updateViewedReviewRequests = async (
 ): Promise<void> => {
   const endpoint = `/sites/${siteName}/review/viewed`
   return apiService.post(endpoint)
-}
-
-export const getSiteLaunchStatus = async (
-  siteName: string
-): Promise<SiteLaunchDto> => {
-  const endpoint = `/sites/${siteName}/launchInfo`
-  return apiService.get(endpoint).then((res) => res.data)
 }
